change log upload path from nvram to /tmp ( OFW-92 )
minor formatting tweaks

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>

drop original scp based version of sync_atom_log_files()

The original version isn't used, so drop it (and then use the
original name for the new tftp based function).

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>

logfiles.sh: rewrite the tftp based sync_atom_log_files()

It's not clear what the advantage of separate "waiting" and
"in progress" lock files is, but it doesn't prevent race conditions
so simplify the code the use a single "in progess" lock file.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>