revert back to setting SYS_DB_FILE to /nvram/syscfg.db in all cases
The recent upstream change to control SYSCFG_DB_FILE based on
BOX_TYPE is not easy to maintain. Hardcode the value to match the
legacy behaviour when the UpdateNvram syscfg value was set to true.

It's not necessary (or safe - since the value may not exist) to test
the UpdateNvram syscfg value.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>