RDKBDEV-1537: drop various obsolete code paths...
Reason for change:
drop various obsolete code paths taken when /etc/device.properties is not found

Testing for /etc/device.properties appears to be an obsolete method
of determining if the image is Yocto based or not. Since all RDKB
builds are Yocto based now (and /etc/device.properties is always
found) the tests and legacy code paths are redundant and can be
removed.

Test Procedure: Sanity.
Risks: None.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
Change-Id: I5f0f795da1070aa48aebc272c953ef77894c24fd