logFiles.properties: enable logfile rotation for Atom CPUInfoPeer.txt.0 ( OFW-5411 )
Fix no log size limit for the Atom side file CPUInfoPeer.txt.0

[Issue Description]
This file can grow up to 1MB or more as there is no limit to it.

[Fix Description]
Included CPUInfoLogFile in File limit size array for atom.
whenever the file size is more than 300KB, the content in the current file is redirect to CPUInfoPeer.txt.0.1
and the current file(CPUInfoPeer.txt.0) is roll over to zero.

[Testing performed]
1. Factory reset the gateway
2. verified the CPUInfoPeer.txt.0 is limited to 300KB.
3. After 300KB, the content is moved to CPUInfoPeer.txt.0.1 and the size of CPUInfoPeer.txt.0 is roll over to OKB

Result:
CPUInfoPeer.txt.0 is limited to 300KB.

Signed-off-by: venkat (vturaka.ext@libertyglobal.com)